main_py/requirements.txt
dotenv
pyjwt

flask
fastapi[standard-no-fastapi-cloud-cli]
psycopg[binary,pool]
pymemcache

langchain
langchain-groq
langchain-google-genai

nomic[local]
numpy
scikit-learn

python-toon
deep-translator

crawl4ai
playwright

holehe
phonenumbers
wappalyzer
pddiktipy

requests>=2.32
urllib3>=2.0
charset-normalizer>=3.0